requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-environ==0.9.0
django-import-export==3.1.0
django-redis==5.2.0
django-rq==2.7.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.5
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
rq==1.13.0
sqlparse==0.4.3
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
backports.zoneinfo==0.2.1;python_version<"3.9"